+++ packages/runtime/src/lib/parsing.ts
@@ -1,5 +1,6 @@
 import { parsing } from "@algo-builder/web";
 import * as base32 from "hi-base32";
+import JSONbig from "json-bigint";
 
 import { RUNTIME_ERRORS } from "../errors/errors-list";
 import { RuntimeError } from "../errors/runtime-errors";
@@ -283,8 +284,7 @@ export function bigEndianBytesToBigInt(bytes: Uint8Array | Buffer): bigint {
 export function strHexToBytes(str: string): Uint8Array {
 	return new Uint8Array(Buffer.from(str.slice(2), "hex"));
 }
-
-/**
+/*
  * Function taken from algosdk.utils
  * ConcatArrays takes n number arrays and returns a joint Uint8Array
  * @param arrs - An arbitrary number of n array-like number list arguments
@@ -301,4 +301,19 @@ export function concatArrays(...arrs: ArrayLike<number>[]) {
 	}
 
 	return c;
+}
+
+/**
+ * assert if given string is a valid JSON object
+ * @param jsonString
+ */
+ export function assertJSON(jsonString: string, line: number): void {
+	const strictBigJSON = JSONbig({ strict: true });
+	try {
+		strictBigJSON.parse(jsonString);
+	} catch (e) {
+		throw new RuntimeError(RUNTIME_ERRORS.TEAL.INVALID_JSON_PARSING, {
+			line: line
+		});
+	}
 }